Add authentication and Facebook integration routes, middleware, and controllers
