+++ .env
@@ -3,3 +3,5 @@ DB_USER=nu_connect_admin
 DB_PASS=$56jisGbjfowKmb&
 DB_NAME=db_nuconnect
 JWT_SECRET=nu_connect_secret_code_natin_to
+FB_ACCESS_TOKEN=EAAQhSZB1ekdoBO2lHTMhZCG85vDHkogy4aMS4izwP7U7AL1K0DzNTgSSFVNhiNNOl0sDssDPkFn7RsXILZBCx5THfwtEQTu1XUpMqbBjjZAyW32TFdCNWFAOeGK49cspZBFp2bk2gbbX7isZC5n16vLdZCZB0kDZAKkoCLkEw0fRgaeBYUncnegFyoQZB2Q8GI9J2C8v0lQHuGqJJFvfZCNPWaz6WO1moZA7bZCkjozNpGZCrs
+FB_GROUP_ID=582726384921248
