+++ config/db.js
@@ -1,0 +1,12 @@
+const mysql = require('mysql2');
+require('dotenv').config();
+
+const con = mysql.createConnection({
+    host: process.env.DB_HOST,
+    user: process.env.DB_USER,
+    password: process.env.DB_PASS,
+    database: process.env.DB_NAME,
+    ssl: { rejectUnauthorized: false }
+});
+
+module.exports = con;
