+++ middlewares/authMiddleware.js
@@ -1,0 +1,21 @@
+const jwt = require('jsonwebtoken');
+require('dotenv').config();
+
+const authMiddleware = (req, res, next) => {
+    const token = req.headers['authorization']?.split(' ')[1]; // Extract token from "Bearer <token>"
+
+    if (!token) {
+        return res.status(401).json({ message: 'No token provided' });
+    }
+
+    jwt.verify(token, process.env.JWT_SECRET, (err, decoded) => {
+        if (err) {
+            return res.status(401).json({ message: 'Invalid token' });
+        }
+
+        req.userId = decoded.id;
+        next();
+    });
+};
+
+module.exports = authMiddleware;
