+++ routes/auth.js
@@ -1,0 +1,7 @@
+const express = require('express');
+const router = express.Router();
+const authController = require('../controllers/authController');
+
+router.post('/login', authController.login);
+
+module.exports = router;
